Maze and Database Expansion
Added the isWinnable method that will check if the maze can be beaten. Returns true. Also integrated the database into the maze, so it can receive questions now. Only thing left I can think of doing for database work is creating an empty source file for the database if none can be found.